Update config_flow.py
Adding "ecotop light with ecomanager-touch" (display) to the config menu, - for users that do not have the Therminator II but the ecotoplight instead. Another option would be to make 2 lines instead of one, that you link internally to the same code... (so it doesn't matter which boiler the user chooses) .. it's just "cosmetics", but reassuring for the user he chose the right thing.